refactor: 컴포넌트 style 분리 및 폴더화
