+++ src/components/EditButton/style.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const StyledButton = styled.button`
+  all: unset;
+  align-self: flex-end;
+  cursor: pointer;
+
+  i {
+    color: ${({ theme }) => theme.color.ink};
+  }
+`;
